refactor: 상금계산 로직 반환타입 int로 수정

+++ src/main/java/domain/stake/Stake.java
@@ -22,7 +22,7 @@ public class Stake {
         }
     }
 
-    public Double getPrize(DealerStatus dealerStatus) {
-        return value * dealerStatus.getMultiply();
+    public int getPrize(DealerStatus dealerStatus) {
+        return (int) (value * dealerStatus.getMultiply());
     }
 }
